refactor(avatar): extracted out AvatarImage
