+++ src/avatar/Avatar.tsx
@@ -2,16 +2,8 @@ import * as React from "react";
 
 import theme from "../theme";
 import { ocx } from "../utils";
-import { useImage } from "../utils/useImage";
-import { GenericAvatar } from "../icons";
+import { AvatarImage } from "./AvatarImage";
 import { useAvatarGroup } from "./AvatarGroup";
-
-function getInitials(name: string) {
-  const [firstName, lastName] = name.split(" ");
-  return firstName && lastName
-    ? `${firstName.charAt(0)}${lastName.charAt(0)}`
-    : firstName.charAt(0);
-}
 
 export type AvatarProps = {
   /**
@@ -74,38 +66,6 @@ export const Avatar: React.FC<AvatarProps> = ({
       />
       {badges}
     </div>
-  );
-};
-
-export type AvatarImageProps = Pick<
-  AvatarProps,
-  "name" | "src" | "onError" | "fallback"
->;
-
-export const AvatarImage: React.FC<AvatarImageProps> = ({
-  src,
-  onError,
-  name,
-  fallback = <GenericAvatar />,
-}) => {
-  const status = useImage({ src, onError });
-
-  const hasLoaded = status === "loaded";
-  const showFallback = !src || !hasLoaded;
-
-  if (showFallback) {
-    return <>{name ? (name ? getInitials?.(name) : null) : fallback}</>;
-  }
-
-  return (
-    <img
-      src={src}
-      alt={name}
-      className={ocx(theme.avatar.image)}
-      style={{
-        borderRadius: "inherit",
-      }}
-    />
   );
 };
